所有者の向きに応じたエフェクトの回転、吸収攻撃が当たらないバグ修正
エフェクト再生時、引数に所有者を設定して向きを取得して再生。

AttackEffectSsytemに向き計算関数追加

バグ：
吸収攻撃判定部分の条件を変更した

+++ Game/Game/source/AttackBase.cpp
@@ -608,9 +608,9 @@ void AttackBase::DrawAttackCollision()
         {
             DrawCircle
             (
-                static_cast<float>(_stcAttackCol.attackColTop.x),   // 中心X座標
-                static_cast<float>(_stcAttackCol.attackColTop.y),   // 中心Y座標
-                _stcAttackCol.attackColR,                           // 半径
+                static_cast<int>(_stcAttackCol.attackColTop.x), // 中心X座標
+                static_cast<int>(_stcAttackCol.attackColTop.y), // 中心Y座標
+                static_cast<int>(_stcAttackCol.attackColR),     // 半径
                 GetColor(0, 255, 0),
                 TRUE
             );
